Added JavaDocs and comments to AbstractBoardGui and Dice classes.
Changed the squares object in AbstractBoardGui from a 1D array to a 2D array of Square objects so the indices now match the corresponding positions in the Game object. This simplifies accessing information from the Game object.

Modified the DicePanel inner class in DiceFrame. It is now a non-static class, and each instance is autonomous, updating itself when its paintComponent() method is called.

Added a new dice image corresponding to the situation where the dice have not yet been rolled (ie there are no scores to display).

Updated the executable Jar file to contain the above changes.